+++ lib/configuration_service.dart
@@ -1,6 +1,5 @@
-import 'package:flutter_background_geolocation/flutter_background_geolocation.dart' as bg;
-
 import 'preferences.dart';
+import 'tracking/engine.dart';
 
 class ConfigurationService {
   static Future<void> applyUri(Uri uri) async {
@@ -23,7 +22,7 @@ class ConfigurationService {
     await _applyBoolParameter(parameters, Preferences.buffer);
     await _applyBoolParameter(parameters, Preferences.wakelock);
     await _applyBoolParameter(parameters, Preferences.stopDetection);
-    await bg.BackgroundGeolocation.setConfig(Preferences.geolocationConfig(true));
+    await engine.setConfig(Preferences.trackingConfig(true));
   }
 
   static Future<void> _applyStringParameter(